move get_csrf_token_from_response to tests utils
