+++ tests/csrf_protection_test.py
@@ -29,20 +29,9 @@ Creation:
 @Date: 2024-05-09
 """
 from bs4 import BeautifulSoup
+from utils import 
 
 
-def get_csrf_token_from_response(html):
-    """Parses the CSRF token from an HTML response.
-
-    Args:
-        html (str): The HTML content from which to parse the CSRF token.
-
-    Returns:
-        str: The CSRF token extracted from the HTML input.
-    """
-    soup = BeautifulSoup(html, "html.parser")
-    token = soup.find("input", {"name": "csrf_token"})["value"]
-    return token
 
 
 def test_login_form_rejection_without_csrf(client):
